# Документацию API (Swagger)/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8090
    description: Generated server url
...
paths:
...
  /user/id:
    get:
      tags:
        - Профиль пользователя
      summary: Get a user by its id
      operationId: get_1
      parameters:
        - name: id
          in: path
          description: id of user to be searched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "404":
          description: User not found
        "200":
          description: Found the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        "400":
          description: Invalid id supplied
    put:
      tags:
        - Профиль пользователя
      operationId: edit_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
    delete:
      tags:
        - Профиль пользователя
      operationId: delete_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
...
  /notification/id:
    get:
      tags:
        - Уведомлени
      summary: Get a notification by its id
      operationId: get_6
      parameters:
        - name: id
          in: path
          description: id of notification to be searched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found the notification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDTO'
        "400":
          description: Invalid id supplied
        "404":
          description: notification not found
    put:
      tags:
        - Уведомлени
      operationId: edit_6
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NotificationDTO'
    delete:
      tags:
        - Уведомлени
      operationId: delete_6
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NotificationDTO'
...
  /carts/{id}:
    get:
      tags:
        - Заказы по клиенту
      operationId: get_10
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CartDto'
    put:
      tags:
        - Заказы по клиенту
      operationId: edit_10
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CartDto'
    delete:
      tags:
        - Заказы по клиенту
      operationId: delete_10
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CartDto'
  /board-preset/id:
    get:
      tags:
        - Доска задач
      summary: Get a board by its id
      operationId: get_11
      parameters:
        - name: id
          in: path
          description: id of board to be searched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found the board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardPresetDTO'
        "400":
          description: Invalid id supplied
        "404":
          description: board not found
    put:
      tags:
        - Доска задач
      operationId: edit_11
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardPresetDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BoardPresetDTO'
    delete:
      tags:
        - Доска задач
      operationId: delete_11
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BoardPresetDTO'
...
  /user:
    get:
      tags:
        - Профиль пользователя
      operationId: getAll_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
    post:
      tags:
        - Профиль пользователя
      operationId: add_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
...
  /notification:
    get:
      tags:
        - Уведомлени
      operationId: getAll_6
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDTO'
    post:
      tags:
        - Уведомлени
      operationId: add_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NotificationDTO'
...
  /carts:
    get:
      tags:
        - Заказы по клиенту
      operationId: getAll_10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartDto'
    post:
      tags:
        - Заказы по клиенту
      operationId: add_10
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CartDto'
  /board-preset:
    get:
      tags:
        - Доска задач
      operationId: getAll_11
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoardPresetDTO'
    post:
      tags:
        - Доска задач
      operationId: add_11
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardPresetDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BoardPresetDTO'
components:
  schemas:
...
    Role:
      required:
        - code
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        code:
          maxLength: 255
          minLength: 0
          type: string
        name:
          maxLength: 255
          minLength: 0
          type: string
        active:
          type: boolean
        new:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        userName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        email:
          type: string
        active:
          type: boolean
        password:
          type: string
        activationCode:
          type: string
        roles:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Role'
        admin:
          type: boolean
        new:
          type: boolean
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        userName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        email:
          type: string
        active:
          type: boolean
        password:
          type: string
        activationCode:
          type: string
        roles:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Role'
    Contractor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        type:
          type: string
          enum:
            - ContractorType()
            - ContractorType()
            - ContractorType()
            - ContractorType()
        email:
          type: array
          items:
            $ref: '#/components/schemas/Email'
        phones:
          type: array
          items:
            $ref: '#/components/schemas/Phone'
        contractorAddress:
          type: array
          items:
            $ref: '#/components/schemas/ContractorAddress'
        new:
          type: boolean
    ContractorAddress:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        description:
          type: string
        address:
          type: string
        fias:
          type: string
        phone:
          type: string
        contactPersone:
          type: string
        new:
          type: boolean
    Email:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        text:
          type: string
        active:
          type: boolean
        new:
          type: boolean
...
    Phone:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        code:
          type: string
        text:
          type: string
        active:
          type: boolean
        new:
          type: boolean
...
    NotificationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        title:
          type: string
        text:
          type: string
        contractor:
          $ref: '#/components/schemas/Contractor'
    OrderItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        qty:
          type: number
          format: double
        plainCartDto:
          $ref: '#/components/schemas/PlainCartDto'
    PlainCartDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    CartDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        itemsDto:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDto'
    BoardPresetDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        text:
          type: string
        boardType:
          type: string
          enum:
            - BoardType()
            - BoardType()
            - BoardType()
        boardView:
          type: string
          enum:
            - BoardView()
            - BoardView()
        filter:
          type: string